Move Unix's webcontext implementation to webkitgtk (#421)
